base: move outcome_future_utils to base library
It needs a home.
