+++ src/v/raft/heartbeat_manager.cc
@@ -10,13 +10,13 @@
 #include "raft/heartbeat_manager.h"
 
 #include "base/likely.h"
+#include "base/outcome_future_utils.h"
 #include "base/vlog.h"
 #include "config/configuration.h"
 #include "features/feature_table.h"
 #include "follower_stats.h"
 #include "model/metadata.h"
 #include "model/timeout_clock.h"
-#include "outcome_future_utils.h"
 #include "raft/consensus_client_protocol.h"
 #include "raft/errc.h"
 #include "raft/group_configuration.h"
