+++ src/v/rpc/include/rpc/connection_set.h
@@ -12,9 +12,9 @@
 #pragma once
 
 #include "base/outcome.h"
+#include "base/outcome_future_utils.h"
 #include "config/tls_config.h"
 #include "model/metadata.h"
-#include "outcome_future_utils.h"
 #include "rpc/backoff_policy.h"
 #include "rpc/errc.h"
 #include "rpc/reconnect_transport.h"
