+++ src/v/utils/tests/outcome_utils_test.cc
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "outcome_future_utils.h"
+#include "base/outcome_future_utils.h"
 
 #include <seastar/testing/thread_test_case.hh>
